Extract Manifest description types into separate files
