+++ Sources/PackageModel/Manifest.swift
@@ -368,399 +368,6 @@ public final class Manifest: ObjectIdentifierProtocol, CustomStringConvertible, 
     }
 }
 
-/// The description of an individual target.
-public struct TargetDescription: Equatable, Codable {
-
-    /// The target type.
-    public enum TargetType: String, Equatable, Codable {
-        case regular
-        case test
-        case system
-        case binary
-    }
-
-    /// Represents a target's dependency on another entity.
-    public enum Dependency: Equatable, ExpressibleByStringLiteral, Codable {
-        case target(name: String, condition: PackageConditionDescription?)
-        case product(name: String, package: String?, condition: PackageConditionDescription?)
-        case byName(name: String, condition: PackageConditionDescription?)
-
-        public init(stringLiteral value: String) {
-            self = .byName(name: value, condition: nil)
-        }
-
-        public static func target(name: String) -> Dependency {
-            return .target(name: name, condition: nil)
-        }
-
-        public static func product(name: String, package: String? = nil) -> Dependency {
-            return .product(name: name, package: package, condition: nil)
-        }
-
-        private enum CodingKeys: String, CodingKey {
-            case target, product, byName
-        }
-
-        public func encode(to encoder: Encoder) throws {
-            var container = encoder.container(keyedBy: CodingKeys.self)
-            switch self {
-            case let .target(a1, a2):
-                var unkeyedContainer = container.nestedUnkeyedContainer(forKey: .target)
-                try unkeyedContainer.encode(a1)
-                try unkeyedContainer.encode(a2)
-            case let .product(a1, a2, a3):
-                var unkeyedContainer = container.nestedUnkeyedContainer(forKey: .product)
-                try unkeyedContainer.encode(a1)
-                try unkeyedContainer.encode(a2)
-                try unkeyedContainer.encode(a3)
-            case let .byName(a1, a2):
-                var unkeyedContainer = container.nestedUnkeyedContainer(forKey: .byName)
-                try unkeyedContainer.encode(a1)
-                try unkeyedContainer.encode(a2)
-            }
-        }
-
-        public init(from decoder: Decoder) throws {
-            let values = try decoder.container(keyedBy: CodingKeys.self)
-            guard let key = values.allKeys.first(where: values.contains) else {
-                throw DecodingError.dataCorrupted(.init(codingPath: decoder.codingPath, debugDescription: "Did not find a matching key"))
-            }
-            switch key {
-            case .target:
-                var unkeyedValues = try values.nestedUnkeyedContainer(forKey: key)
-                let a1 = try unkeyedValues.decode(String.self)
-                let a2 = try unkeyedValues.decodeIfPresent(PackageConditionDescription.self)
-                self = .target(name: a1, condition: a2)
-            case .product:
-                var unkeyedValues = try values.nestedUnkeyedContainer(forKey: key)
-                let a1 = try unkeyedValues.decode(String.self)
-                let a2 = try unkeyedValues.decodeIfPresent(String.self)
-                let a3 = try unkeyedValues.decodeIfPresent(PackageConditionDescription.self)
-                self = .product(name: a1, package: a2, condition: a3)
-            case .byName:
-                var unkeyedValues = try values.nestedUnkeyedContainer(forKey: key)
-                let a1 = try unkeyedValues.decode(String.self)
-                let a2 = try unkeyedValues.decodeIfPresent(PackageConditionDescription.self)
-                self = .byName(name: a1, condition: a2)
-            }
-        }
-    }
-
-    public struct Resource: Codable, Equatable {
-        public enum Rule: String, Codable, Equatable {
-            case process
-            case copy
-        }
-
-        public enum Localization: String, Codable, Equatable {
-            case `default`
-            case base
-        }
-
-        /// The rule for the resource.
-        public let rule: Rule
-
-        /// The path of the resource.
-        public let path: String
-
-        /// The explicit localization of the resource.
-        public let localization: Localization?
-
-        public init(rule: Rule, path: String, localization: Localization? = nil) {
-            precondition(rule == .process || localization == nil)
-            self.rule = rule
-            self.path = path
-            self.localization = localization
-        }
-    }
-
-    /// The name of the target.
-    public let name: String
-
-    /// The custom path of the target.
-    public let path: String?
-
-    /// The url of the binary target artifact.
-    public let url: String?
-
-    /// The custom sources of the target.
-    public let sources: [String]?
-
-    /// The explicitly declared resources of the target.
-    public let resources: [Resource]
-
-    /// The exclude patterns.
-    public let exclude: [String]
-
-    // FIXME: Kill this.
-    //
-    /// Returns true if the target type is test.
-    public var isTest: Bool {
-        return type == .test
-    }
-
-    /// The declared target dependencies.
-    public let dependencies: [Dependency]
-
-    /// The custom public headers path.
-    public let publicHeadersPath: String?
-
-    /// The type of target.
-    public let type: TargetType
-
-    /// The pkg-config name of a system library target.
-    public let pkgConfig: String?
-
-    /// The providers of a system library target.
-    public let providers: [SystemPackageProviderDescription]?
-
-    /// The target-specific build settings declared in this target.
-    public let settings: [TargetBuildSettingDescription.Setting]
-
-    /// The binary target checksum.
-    public let checksum: String?
-
-    public init(
-        name: String,
-        dependencies: [Dependency] = [],
-        path: String? = nil,
-        url: String? = nil,
-        exclude: [String] = [],
-        sources: [String]? = nil,
-        resources: [Resource] = [],
-        publicHeadersPath: String? = nil,
-        type: TargetType = .regular,
-        pkgConfig: String? = nil,
-        providers: [SystemPackageProviderDescription]? = nil,
-        settings: [TargetBuildSettingDescription.Setting] = [],
-        checksum: String? = nil
-    ) {
-        switch type {
-        case .regular, .test:
-            precondition(
-                url == nil &&
-                pkgConfig == nil &&
-                providers == nil &&
-                checksum == nil
-            )
-        case .system:
-            precondition(
-                dependencies.isEmpty &&
-                exclude.isEmpty &&
-                sources == nil &&
-                resources.isEmpty &&
-                publicHeadersPath == nil &&
-                settings.isEmpty &&
-                checksum == nil
-            )
-        case .binary:
-            precondition(path != nil || url != nil)
-            precondition(
-                dependencies.isEmpty &&
-                exclude.isEmpty &&
-                sources == nil &&
-                resources.isEmpty &&
-                publicHeadersPath == nil &&
-                pkgConfig == nil &&
-                providers == nil &&
-                settings.isEmpty
-            )
-        }
-
-        self.name = name
-        self.dependencies = dependencies
-        self.path = path
-        self.url = url
-        self.publicHeadersPath = publicHeadersPath
-        self.sources = sources
-        self.exclude = exclude
-        self.resources = resources
-        self.type = type
-        self.pkgConfig = pkgConfig
-        self.providers = providers
-        self.settings = settings
-        self.checksum = checksum
-    }
-}
-
-/// The product description
-public struct ProductDescription: Equatable, Codable {
-
-    /// The name of the product.
-    public let name: String
-
-    /// The targets in the product.
-    public let targets: [String]
-
-    /// The type of product.
-    public let type: ProductType
-
-    public init(
-        name: String,
-        type: ProductType,
-        targets: [String]
-    ) {
-        precondition(type != .test, "Declaring test products isn't supported: \(name):\(targets)")
-        self.name = name
-        self.type = type
-        self.targets = targets
-    }
-}
-
-/// Represents system package providers.
-public enum SystemPackageProviderDescription: Equatable, Codable {
-    case brew([String])
-    case apt([String])
-    case yum([String])
-
-    private enum CodingKeys: String, CodingKey {
-        case brew, apt, yum
-    }
-
-    public func encode(to encoder: Encoder) throws {
-        var container = encoder.container(keyedBy: CodingKeys.self)
-        switch self {
-        case let .brew(a1):
-            var unkeyedContainer = container.nestedUnkeyedContainer(forKey: .brew)
-            try unkeyedContainer.encode(a1)
-        case let .apt(a1):
-            var unkeyedContainer = container.nestedUnkeyedContainer(forKey: .apt)
-            try unkeyedContainer.encode(a1)
-        case let .yum(a1):
-            var unkeyedContainer = container.nestedUnkeyedContainer(forKey: .yum)
-            try unkeyedContainer.encode(a1)
-        }
-    }
-
-    public init(from decoder: Decoder) throws {
-        let values = try decoder.container(keyedBy: CodingKeys.self)
-        guard let key = values.allKeys.first(where: values.contains) else {
-            throw DecodingError.dataCorrupted(.init(codingPath: decoder.codingPath, debugDescription: "Did not find a matching key"))
-        }
-        switch key {
-        case .brew:
-            var unkeyedValues = try values.nestedUnkeyedContainer(forKey: key)
-            let a1 = try unkeyedValues.decode([String].self)
-            self = .brew(a1)
-        case .apt:
-            var unkeyedValues = try values.nestedUnkeyedContainer(forKey: key)
-            let a1 = try unkeyedValues.decode([String].self)
-            self = .apt(a1)
-        case .yum:
-            var unkeyedValues = try values.nestedUnkeyedContainer(forKey: key)
-            let a1 = try unkeyedValues.decode([String].self)
-            self = .yum(a1)
-        }
-    }
-}
-
-/// Represents a package dependency.
-public struct PackageDependencyDescription: Equatable, Codable {
-
-    /// The dependency requirement.
-    public enum Requirement: Equatable, Hashable, CustomStringConvertible, Codable {
-        case exact(Version)
-        case range(Range<Version>)
-        case revision(String)
-        case branch(String)
-        case localPackage
-
-        public static func upToNextMajor(from version: TSCUtility.Version) -> Requirement {
-            return .range(version..<Version(version.major + 1, 0, 0))
-        }
-
-        public static func upToNextMinor(from version: TSCUtility.Version) -> Requirement {
-            return .range(version..<Version(version.major, version.minor + 1, 0))
-        }
-
-        public var description: String {
-            switch self {
-            case .exact(let version):
-                return version.description
-            case .range(let range):
-                return range.description
-            case .revision(let revision):
-                return "revision[\(revision)]"
-            case .branch(let branch):
-                return "branch[\(branch)]"
-            case .localPackage:
-                return "local"
-            }
-        }
-
-        private enum CodingKeys: String, CodingKey {
-            case exact, range, revision, branch, localPackage
-        }
-
-        public func encode(to encoder: Encoder) throws {
-            var container = encoder.container(keyedBy: CodingKeys.self)
-            switch self {
-            case let .exact(a1):
-                var unkeyedContainer = container.nestedUnkeyedContainer(forKey: .exact)
-                try unkeyedContainer.encode(a1)
-            case let .range(a1):
-                var unkeyedContainer = container.nestedUnkeyedContainer(forKey: .range)
-                try unkeyedContainer.encode(CodableRange(a1))
-            case let .revision(a1):
-                var unkeyedContainer = container.nestedUnkeyedContainer(forKey: .revision)
-                try unkeyedContainer.encode(a1)
-            case let .branch(a1):
-                var unkeyedContainer = container.nestedUnkeyedContainer(forKey: .branch)
-                try unkeyedContainer.encode(a1)
-            case .localPackage:
-                try container.encodeNil(forKey: .localPackage)
-            }
-        }
-
-        public init(from decoder: Decoder) throws {
-            let values = try decoder.container(keyedBy: CodingKeys.self)
-            guard let key = values.allKeys.first(where: values.contains) else {
-                throw DecodingError.dataCorrupted(.init(codingPath: decoder.codingPath, debugDescription: "Did not find a matching key"))
-            }
-            switch key {
-            case .exact:
-                var unkeyedValues = try values.nestedUnkeyedContainer(forKey: key)
-                let a1 = try unkeyedValues.decode(Version.self)
-                self = .exact(a1)
-            case .range:
-                var unkeyedValues = try values.nestedUnkeyedContainer(forKey: key)
-                let a1 = try unkeyedValues.decode(CodableRange<Version>.self)
-                self = .range(a1.range)
-            case .revision:
-                var unkeyedValues = try values.nestedUnkeyedContainer(forKey: key)
-                let a1 = try unkeyedValues.decode(String.self)
-                self = .revision(a1)
-            case .branch:
-                var unkeyedValues = try values.nestedUnkeyedContainer(forKey: key)
-                let a1 = try unkeyedValues.decode(String.self)
-                self = .branch(a1)
-            case .localPackage:
-                self = .localPackage
-            }
-        }
-    }
-
-    /// The name of the dependency explicitly defined in the manifest.
-    public let explicitName: String?
-
-    /// The name of the dependency, either explicitly defined in the manifest, or deduced from the URL.
-    public let name: String
-
-    /// The url of the dependency.
-    public let url: String
-
-    /// The dependency requirement.
-    public let requirement: Requirement
-
-    /// Create a dependency.
-    public init(name: String?, url: String, requirement: Requirement) {
-        self.explicitName = name
-        self.name = name ?? PackageReference.computeDefaultName(fromURL: url)
-        self.url = url
-        self.requirement = requirement
-    }
-}
-
 /// The products requested of a package.
 ///
 /// Any product which matches the filter will be used for dependency resolution, whereas unrequested products will be ingored.
@@ -846,174 +453,5 @@ public enum ProductFilter: Equatable, Hashable, Codable, JSONMappable, JSONSeria
         case .specific(let set):
             return "[\(set.sorted().joined(separator: ", "))]"
         }
-    }
-}
-
-/// A dependency description along with its associated product filter.
-public struct FilteredDependencyDescription: Codable {
-
-    /// Creates a filtered dependency.
-    ///
-    /// - Parameters:
-    ///   - declaration: The raw dependency.
-    ///   - productFilter: The product filter to apply.
-    public init(declaration: PackageDependencyDescription, productFilter: ProductFilter) {
-        self.declaration = declaration
-        self.productFilter = productFilter
-    }
-
-    /// The loaded dependency declaration.
-    public let declaration: PackageDependencyDescription
-
-    /// The resolved product filter.
-    public let productFilter: ProductFilter
-}
-
-public struct PlatformDescription: Codable, Equatable {
-    public let platformName: String
-    public let version: String
-    public let options: [String]
-
-    public init(name: String, version: String, options: [String] = []) {
-        self.platformName = name
-        self.version = version
-        self.options = options
-    }
-}
-
-/// Represents a manifest condition.
-public struct PackageConditionDescription: Codable, Equatable {
-    public let platformNames: [String]
-    public let config: String?
-
-    public init(platformNames: [String] = [], config: String? = nil) {
-        assert(!(platformNames.isEmpty && config == nil))
-        self.platformNames = platformNames
-        self.config = config
-    }
-}
-
-/// A namespace for target-specific build settings.
-public enum TargetBuildSettingDescription {
-
-    /// The tool for which a build setting is declared.
-    public enum Tool: String, Codable, Equatable, CaseIterable {
-        case c
-        case cxx
-        case swift
-        case linker
-    }
-
-    /// The name of the build setting.
-    public enum SettingName: String, Codable, Equatable {
-        case headerSearchPath
-        case define
-        case linkedLibrary
-        case linkedFramework
-
-        case unsafeFlags
-    }
-
-    /// An individual build setting.
-    public struct Setting: Codable, Equatable {
-
-        /// The tool associated with this setting.
-        public let tool: Tool
-
-        /// The name of the setting.
-        public let name: SettingName
-
-        /// The condition at which the setting should be applied.
-        public let condition: PackageConditionDescription?
-
-        /// The value of the setting.
-        ///
-        /// This is kind of like an "untyped" value since the length
-        /// of the array will depend on the setting type.
-        public let value: [String]
-
-        public init(
-            tool: Tool,
-            name: SettingName,
-            value: [String],
-            condition: PackageConditionDescription? = nil
-        ) {
-            switch name {
-            case .headerSearchPath: fallthrough
-            case .define: fallthrough
-            case .linkedLibrary: fallthrough
-            case .linkedFramework:
-                assert(value.count == 1, "\(tool) \(name) \(value)")
-                break
-            case .unsafeFlags:
-                assert(value.count >= 1, "\(tool) \(name) \(value)")
-                break
-            }
-
-            self.tool = tool
-            self.name = name
-            self.value = value
-            self.condition = condition
-        }
-    }
-}
-
-/// A manifest condition.
-public protocol PackageConditionProtocol: Codable {
-    func satisfies(_ environment: BuildEnvironment) -> Bool
-}
-
-/// Wrapper for package condition so it can be conformed to Codable.
-struct PackageConditionWrapper: Codable {
-    var platform: PlatformsCondition?
-    var config: ConfigurationCondition?
-
-    var condition: PackageConditionProtocol {
-        if let platform = platform {
-            return platform
-        } else if let config = config {
-            return config
-        } else {
-            fatalError("unreachable")
-        }
-    }
-
-    init(_ condition: PackageConditionProtocol) {
-        switch condition {
-        case let platform as PlatformsCondition:
-            self.platform = platform
-        case let config as ConfigurationCondition:
-            self.config = config
-        default:
-            fatalError("unknown condition \(condition)")
-        }
-    }
-}
-
-/// Platforms condition implies that an assignment is valid on these platforms.
-public struct PlatformsCondition: PackageConditionProtocol {
-    public let platforms: [Platform]
-
-    public init(platforms: [Platform]) {
-        assert(!platforms.isEmpty, "List of platforms should not be empty")
-        self.platforms = platforms
-    }
-
-    public func satisfies(_ environment: BuildEnvironment) -> Bool {
-        platforms.contains(environment.platform)
-    }
-}
-
-/// A configuration condition implies that an assignment is valid on
-/// a particular build configuration.
-public struct ConfigurationCondition: PackageConditionProtocol {
-    public let configuration: BuildConfiguration
-
-    public init(configuration: BuildConfiguration) {
-        self.configuration = configuration
-    }
-
-    public func satisfies(_ environment: BuildEnvironment) -> Bool {
-        configuration == environment.configuration
     }
 }
