+++ Sources/PackageModel/Manifest/PackageDependencyDescription.swift
@@ -1,0 +1,118 @@
+/*
+ This source file is part of the Swift.org open source project
+
+ Copyright (c) 2014 - 2020 Apple Inc. and the Swift project authors
+ Licensed under Apache License v2.0 with Runtime Library Exception
+
+ See http://swift.org/LICENSE.txt for license information
+ See http://swift.org/CONTRIBUTORS.txt for Swift project authors
+*/
+
+import TSCBasic
+
+/// Represents a package dependency.
+public struct PackageDependencyDescription: Equatable, Codable {
+
+    /// The dependency requirement.
+    public enum Requirement: Equatable, Hashable, CustomStringConvertible, Codable {
+        case exact(Version)
+        case range(Range<Version>)
+        case revision(String)
+        case branch(String)
+        case localPackage
+
+        public static func upToNextMajor(from version: TSCUtility.Version) -> Requirement {
+            return .range(version..<Version(version.major + 1, 0, 0))
+        }
+
+        public static func upToNextMinor(from version: TSCUtility.Version) -> Requirement {
+            return .range(version..<Version(version.major, version.minor + 1, 0))
+        }
+
+        public var description: String {
+            switch self {
+            case .exact(let version):
+                return version.description
+            case .range(let range):
+                return range.description
+            case .revision(let revision):
+                return "revision[\(revision)]"
+            case .branch(let branch):
+                return "branch[\(branch)]"
+            case .localPackage:
+                return "local"
+            }
+        }
+
+        private enum CodingKeys: String, CodingKey {
+            case exact, range, revision, branch, localPackage
+        }
+
+        public func encode(to encoder: Encoder) throws {
+            var container = encoder.container(keyedBy: CodingKeys.self)
+            switch self {
+            case let .exact(a1):
+                var unkeyedContainer = container.nestedUnkeyedContainer(forKey: .exact)
+                try unkeyedContainer.encode(a1)
+            case let .range(a1):
+                var unkeyedContainer = container.nestedUnkeyedContainer(forKey: .range)
+                try unkeyedContainer.encode(CodableRange(a1))
+            case let .revision(a1):
+                var unkeyedContainer = container.nestedUnkeyedContainer(forKey: .revision)
+                try unkeyedContainer.encode(a1)
+            case let .branch(a1):
+                var unkeyedContainer = container.nestedUnkeyedContainer(forKey: .branch)
+                try unkeyedContainer.encode(a1)
+            case .localPackage:
+                try container.encodeNil(forKey: .localPackage)
+            }
+        }
+
+        public init(from decoder: Decoder) throws {
+            let values = try decoder.container(keyedBy: CodingKeys.self)
+            guard let key = values.allKeys.first(where: values.contains) else {
+                throw DecodingError.dataCorrupted(.init(codingPath: decoder.codingPath, debugDescription: "Did not find a matching key"))
+            }
+            switch key {
+            case .exact:
+                var unkeyedValues = try values.nestedUnkeyedContainer(forKey: key)
+                let a1 = try unkeyedValues.decode(Version.self)
+                self = .exact(a1)
+            case .range:
+                var unkeyedValues = try values.nestedUnkeyedContainer(forKey: key)
+                let a1 = try unkeyedValues.decode(CodableRange<Version>.self)
+                self = .range(a1.range)
+            case .revision:
+                var unkeyedValues = try values.nestedUnkeyedContainer(forKey: key)
+                let a1 = try unkeyedValues.decode(String.self)
+                self = .revision(a1)
+            case .branch:
+                var unkeyedValues = try values.nestedUnkeyedContainer(forKey: key)
+                let a1 = try unkeyedValues.decode(String.self)
+                self = .branch(a1)
+            case .localPackage:
+                self = .localPackage
+            }
+        }
+    }
+
+    /// The name of the dependency explicitly defined in the manifest.
+    public let explicitName: String?
+
+    /// The name of the dependency, either explicitly defined in the manifest, or deduced from the URL.
+    public let name: String
+
+    /// The url of the dependency.
+    public let url: String
+
+    /// The dependency requirement.
+    public let requirement: Requirement
+
+    /// Create a dependency.
+    public init(name: String?, url: String, requirement: Requirement) {
+        self.explicitName = name
+        self.name = name ?? PackageReference.computeDefaultName(fromURL: url)
+        self.url = url
+        self.requirement = requirement
+    }
+}
